[Fix] Moved Shared Constants To New File (#277)
* Created shared constants script and added two Enums to the file

* Adding MessageKeysEnum as shared, removed original and changing all related imports

* Adding TestStateEnum and changing all relevant imports

* Exchanging for a app relative import in the models' init file

* Updating Copyright's Date Year

* Reverting python_tests SHA since it's not ready yet

* Updating Matter test scripts SHA
