+++ app/constants/websockets_constants.py
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2023 Project CHIP Authors
+# Copyright (c) 2025 Project CHIP Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -27,23 +27,6 @@ UDP_SOCKET_PORT = 5000
 UDP_SOCKET_INTERFACE = "0.0.0.0"
 
 
-# Enum Keys for different types of messages currently supported by the tool
-class MessageTypeEnum(str, Enum):
-    PROMPT_REQUEST = "prompt_request"
-    OPTIONS_REQUEST = "options_request"
-    MESSAGE_REQUEST = "message_request"
-    FILE_UPLOAD_REQUEST = "file_upload_request"
-    PROMPT_RESPONSE = "prompt_response"
-    TEST_UPDATE = "test_update"
-    TIME_OUT_NOTIFICATION = "time_out_notification"
-    TEST_LOG_RECORDS = "test_log_records"
-    INVALID_MESSAGE = "invalid_message"
-    STREAM_VERIFICATION_REQUEST = "stream_verification_request"
-    IMAGE_VERIFICATION_REQUEST = "image_verification_request"
-    TWO_WAY_TALK_VERIFICATION_REQUEST = "two_way_talk_verification_request"
-    PUSH_AV_STREAM_VERIFICATION_REQUEST = "push_av_stream_verification_request"
-
-
 class WebSocketTypeEnum(str, Enum):
     MAIN = "main"
     VIDEO = "video"
@@ -55,9 +38,3 @@ class WebSocketConnection:
     def __init__(self, websocket: WebSocket, socket_type: WebSocketTypeEnum) -> None:
         self.websocket = websocket
         self.type = socket_type
-
-
-# Enum keys used with messages at the top level
-class MessageKeysEnum(str, Enum):
-    TYPE = "type"
-    PAYLOAD = "payload"
