+++ app/models/__init__.py
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2023 Project CHIP Authors
+# Copyright (c) 2025 Project CHIP Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,11 +13,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+from app.constants.shared_constants import TestStateEnum
+
 from .operator import Operator
 from .project import Project
 from .test_case_execution import TestCaseExecution
 from .test_case_metadata import TestCaseMetadata
-from .test_enums import TestStateEnum
 from .test_run_config import TestRunConfig
 from .test_run_execution import TestRunExecution
 from .test_step_execution import TestStepExecution
