--- app/models/test_enums.py
@@ -1,28 +1,0 @@
-#
-# Copyright (c) 2023 Project CHIP Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-from enum import Enum
-
-
-class TestStateEnum(str, Enum):
-    __test__ = False  # Needed to indicate to PyTest that this is not a "test"
-    PENDING = "pending"
-    EXECUTING = "executing"
-    PENDING_ACTUATION = "pending_actuation"  # TODO: Do we need this
-    PASSED = "passed"  # Test Passed with no issued
-    FAILED = "failed"  # Test Failed
-    ERROR = "error"  # Test Error due to tool setup or environment
-    NOT_APPLICABLE = "not_applicable"  # Test is not applicable - e.g. PICS mismatch
-    CANCELLED = "cancelled"
