+++ app/test_engine/models/test_step.py
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2023 Project CHIP Authors
+# Copyright (c) 2025 Project CHIP Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -15,7 +15,7 @@
 #
 from typing import List, Optional
 
-from app.models.test_enums import TestStateEnum
+from app.models import TestStateEnum
 from app.models.test_step_execution import TestStepExecution
 from app.test_engine.logger import test_engine_logger as logger
 from app.test_engine.models.utils import LogSeparator
