+++ app/tests/crud/test_test_run_execution.py
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2023 Project CHIP Authors
+# Copyright (c) 2025 Project CHIP Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -27,7 +27,7 @@ from sqlalchemy.orm import Session
 
 from app import crud, models, schemas
 from app.crud.crud_test_run_execution import ImportError
-from app.models.test_enums import TestStateEnum
+from app.models import TestStateEnum
 from app.schemas.test_run_config import TestRunConfigCreate
 from app.schemas.test_run_execution import (
     TestRunExecutionCreate,
