+++ app/tests/test_engine/test_ui_observer.py
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2023 Project CHIP Authors
+# Copyright (c) 2025 Project CHIP Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -20,8 +20,11 @@ from unittest import mock
 import pytest
 from sqlalchemy.orm import Session
 
-from app.constants.websockets_constants import MessageKeysEnum, MessageTypeEnum
-from app.models.test_enums import TestStateEnum
+from app.constants.shared_constants import (
+    MessageKeysEnum,
+    MessageTypeEnum,
+    TestStateEnum,
+)
 from app.models.test_run_execution import TestRunExecution
 from app.schemas.test_run_log_entry import TestRunLogEntry
 from app.test_engine.models import TestRun
