+++ app/tests/user_prompts/test_user_prompt_manager.py
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2023 Project CHIP Authors
+# Copyright (c) 2025 Project CHIP Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -17,11 +17,8 @@ from unittest import mock
 
 import pytest
 
-from app.constants.websockets_constants import (
-    MESSAGE_ID_KEY,
-    MessageKeysEnum,
-    MessageTypeEnum,
-)
+from app.constants.shared_constants import MessageKeysEnum, MessageTypeEnum
+from app.constants.websockets_constants import MESSAGE_ID_KEY
 from app.user_prompt_support.user_prompt_manager import (
     PromptExchange,
     user_prompt_manager,
