+++ app/tests/utils/test_step_execution.py
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2023 Project CHIP Authors
+# Copyright (c) 2025 Project CHIP Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -18,7 +18,7 @@ from typing import Any, Dict, List, Optional
 
 from faker import Faker
 
-from app.models.test_enums import TestStateEnum
+from app.models import TestStateEnum
 
 fake = Faker()
