+++ test_collections/matter/sdk_tests/support/python_testing/models/utils.py
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2023-2024 Project CHIP Authors
+# Copyright (c) 2023-2025 Project CHIP Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -21,6 +21,7 @@ from typing import Generator, cast
 
 import loguru
 
+from app.constants.shared_constants import DutPairingModeEnum
 from app.schemas.test_environment_config import ThreadAutoConfig
 from app.test_engine.logger import PYTHON_TEST_LEVEL
 from app.user_prompt_support import UserPromptSupport
@@ -28,7 +29,6 @@ from test_collections.matter.sdk_tests.support.otbr_manager.otbr_manager import 
     ThreadBorderRouter,
 )
 from test_collections.matter.test_environment_config import (
-    DutPairingModeEnum,
     TestEnvironmentConfigMatter,
     ThreadExternalConfig,
 )
